+++ packages/react/src/atomic/Divider/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Dividers are used to separate content in a list or grid. They can be horizontal or vertical.
 
+## 2026-03-02 `1.1.0`
+
+#### 🎉 New features
+
+- Added `width`, `marginLeft` and `marginRight` props to Divider component. ([#437](https://github.com/TiendaNube/nimbus-design-system/pull/437) by [@joacotornello](https://github.com/joacotornello))
+
 ## 2026-01-13 `1.0.1`
 
 #### 🎉 New features